working

+++ src/index_main.cpp
@@ -182,7 +182,7 @@ Index build_index(const InputFileMap& input, InputSummary& summary, const IndexA
                 ibf.emplace(value, seqan3::bin_index{bin});
             record_count++;
         }
-        PLOG_INFO << "Added file " << fasta_file << " with " << record_count << " records to bin " << bin << std::endl;
+        PLOG_INFO << "Added file " << fasta_file << " with " << record_count << " records to bin " << static_cast<uint16_t>(bin) << std::endl;
     }
 
     return Index(opt, ibf);
